Rename treasury transaction module
